+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New curves classes: Parabola2D/3D.
 - ConicalSurface3D: line/line_segment intersections, perpendicular_plane_intersection, parallel_plane_intersections, concurent_plane_intersections, plane_intersections.
 - Hyperbola2D/3D and Parabola2D/3D: split
-- 
+- PlaneFace3D: conicalface_intersections
+
 ### Fixed
 - add missing name attributes to classmethods
 - fixed circular imports